Cleanup functions in class defn
