+++ src/utils/setup/get-relevant-snooty-node-content.js
@@ -1,0 +1,42 @@
+/**
+ * Name map of directives we want to display in an article
+ */
+const contentNodesMap = {
+    introduction: true,
+    prerequisites: true,
+    content: true,
+    summary: true,
+};
+
+/**
+ * search the ast for the few directives we need to display content
+ * @param {array} nodes
+ * @returns {array} array of childNodes with our main content
+ */
+export const getRelevantSnootyNodeContent = nodes => {
+    const nodesWeActuallyWant = [];
+    for (let nodeIndex = 0; nodeIndex < nodes.length; nodeIndex++) {
+        const childNode = nodes[nodeIndex];
+        // check top level directives first
+        if (contentNodesMap[childNode.name]) {
+            //@ts-ignore
+            nodesWeActuallyWant.push(childNode);
+        }
+        // Some content nodes will be children of section nodes
+        else if (childNode.type === 'section') {
+            for (
+                let childIndex = 0;
+                childIndex < childNode.children.length;
+                childIndex++
+            ) {
+                const grandChildNode = childNode.children[childIndex];
+                if (contentNodesMap[grandChildNode.name]) {
+                    //@ts-ignore
+                    nodesWeActuallyWant.push(grandChildNode);
+                }
+            }
+        }
+    }
+
+    return nodesWeActuallyWant;
+};
